Split combined forms in to three speperate form components, test, and plug into parents
